add a error page with navigation
